### 2016-10-03 #78 @perrystephenson
refimpact submission
Provides wrapper functions around the UK Research Excellence Framework 2014 Impact Case Studies Database API http://impact.ref.ac.uk/. The database contains relevant publication and research metadata about each case study as well as several paragraphs of text from the case study submissions. Case studies in the database are licenced under a CC-BY 4.0 licence http://creativecommons.org/licenses/by/4.0/legalcode.

```
Package: refimpact
Title: API Wrapper for the UK REF 2014 Impact Case Studies Database
Version: 0.1.0
Authors@R: person("Perry", "Stephenson",
    email = "perry.stephenson+cran@gmail.com",
    role = c("aut", "cre"))
Description: Provides wrapper functions around the UK Research
    Excellence Framework 2014 Impact Case Studies Database API
    <http://impact.ref.ac.uk/>. The database contains relevant publication and
    research metadata about each case study as well as several paragraphs of
    text from the case study submissions. Case studies in the database are
    licenced under a CC-BY 4.0 licence
    <http://creativecommons.org/licenses/by/4.0/legalcode>.
    curl (>= 2.1),
    jsonlite (>= 1.1),
    tibble (>= 1.2)
Suggests: testthat
```

https://github.com/perrystephenson/refimpact

People interested in the following:
1. Research in the UK and it's global impact
2. Relationships between academic institutions
3. Relationships between impactful publications
4. A high quality and homogenous text resource for text mining

There are no other packages for accessing this data, as far as I am aware.

- [ ] contains a vignette with examples of its essential functions and uses. (this is currently in the README.md file, but I can add a vignette if required)
- [X] Do you intend for this package to go on CRAN?  (Already accepted - [link](https://cran.r-project.org/package=refimpact))
  - (_Do not submit your package separately to JOSS_)
- [X] Does `R CMD check` (or `devtools::check()`) succeed?  Paste and describe any errors or warnings: No errors, warnings nor notes.
