### 2016-10-03 #78 @sckott
### Editor checks:

- [ ] **Fit**: The package meets or criteria [fit](https://github.com/ropensci/onboarding/blob/master/policies.md#package-fit) and [overlap](https://github.com/ropensci/onboarding/blob/master/policies.md#package-overlap)

---

#### Editor comments

Thanks for your submission @perrystephenson

In assessing fit, I have a few questions:

* How widely used is this data in research/academia/non-profits?  From a [google scholar search](https://scholar.google.com/scholar?hl=en&q=%22research+excellence+framework%22&btnG=&as_sdt=1%2C38) it seems to be mentioned/cited a lot.
* Do you know if this data is meant to be temporarily available, or meant to be around for a while?
* Are there similar things in other countries?  (just curiosity)
* You mentioned text mining. I didn't see egs. in your docs though. What kind of text is available through the service?
