### 2016-10-05 #78 @perrystephenson
Hi @sckott

Thanks for getting back to me so quickly! I'll reply in order:

1. It has been used predominantly in text mining research projects, an example is [here](http://www.hefce.ac.uk/pubs/rereports/Year/2015/analysisREFimpact/). We're also working on it at the University of Technology, Sydney (UTS) with a view to building tools which can help researchers improve the quality of their academic writing by making it "more like" an existing body of known high quality text. This is related to our work on the [Academic Writing Analysis](https://utscic.edu.au/tools/awa/) project.
2. I have looked around previously for an answer to this question, and everything I have read seems to imply that the data will be around for a while. I suspect that they will build up the dataset over time, which will mean adding the 2019 submission data when it becomes available.
3. I am not aware of similar datasets in other countries (unfortunately). If my project proves successful then perhaps UTS will lobby the Australian Research Council to release a similar dataset.
4. The text available through the service is the full text of over 6,000 research impact case study submissions, as well as a few structured and semi-structured pieces of metadata. This is a large collection of similar documents, with consistent grammar and minimal spelling errors. I am planning to write a vignette for the next CRAN release, which will incorporate some of my work on the project mentioned in point 1.

Thanks!
