### 2016-10-07 #78 @sckott
@perrystephenson checked the fit box ✔️

Seeking reviewers now, see updated comment above about `goodpractice`
