### 2016-10-19 #78 @ropenscibot
@njahn82 @kbenoit - hey there, it's been 16 days, please get your review in by Oct 24, thanks :smiley_cat: (ropensci-bot)
