### 2016-10-19 #78 @sckott
@njahn82 @kbenoit  sorry about the ping from the bot (message deleted, but you probably got an email). still working out the rough spots
