### 2016-10-21 #78 @ropenscibot
@njahn82 @kbenoit - hey there, it's been 18 days, please get your review in by Oct 24, thanks :smiley_cat: (ropensci-bot)
