### 2016-10-31 #78 @njahn82
>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly staing problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This package gives access to UK's 2014 Research Excellence Framework (REF) Impact Case Studies. To my knowledge, this is one of the first comprehensive efforts to share reports under an open content license, making it an interesting source for studying science, science communication and funding activities.

The package interfaces the REF2014 API. Although the interfaced database seems to contain work being completed, obtaining data from the API rather than as package data is very reasonable because CRAN limits the size of  data files to reduce installation time. Besides, the API provides various query methods, which are accessible through the package. The main function of `refimpact` is `get_case_studies()`

The author did a nice job interfacing the API with `jsonlite` and documenting the functions. Using `tibble` is a great choice to encourage re-use of the fetched data within R. The package is also available via CRAN, so it has passed essential tests when sharing an R package. However, I feel that the package could benefit from some revisions in order to increase code quality and rOpenSci compliance.

#### Use httr to work with and manipulate http requests

I would strongly recommend using [httr](https://cran.r-project.org/web/packages/httr/index.html) for building and accessing data from web APIs. `httr` provides useful functions to create API queries, and increases the reliability of packages when something breaks. [This httr vignette](https://cran.r-project.org/web/packages/httr/vignettes/api-packages.html) is a great introduction about how to write an API package.

More specifically:

- Input validation and building the queries can be done with `httr:::modify_url()`. This would reduce the current code used and thus make it more reliable.

- APIs can break. `httr` allows converting http errors to R, and testing if content from a request is valid json. Both methods should be implemented in `refimpact` to make it more transparent to the users when a request fails.

#### Provide a vignette

The package would benefit from a [vignette, a long-form documentation of how to use the package](http://r-pkgs.had.co.nz/vignettes.html). The README nicely summarises who might be interested in using `refimpact`, how to install and how to access the data. This could be a great start and continued with a demonstration of how to solve particular problems researchers are faced with when they study science and its funding. In particular, it would be nice to see how to work with `refimpact` output using other R packages written for text-mining purposes or network analysis. I also see a lot of potential when interfacing `refimpact` data with literature packages to get more detailed bibliographic records.

* Please follow rOpenSci guidelines and create the `README.md` directly from a `README.Rmd`. One advantage is that you can present both code and results to your audience.

#### Functions

* Instead of integers, I would suggest to use character vectors as input for `get_case_studies()`. Although REF's variables are syntactically numeric, they should be handled as categorical variables.
* I wonder if the parameter `phrase` in `get_case_studies()` could be improved to allow more than a single word as input?
* There is some potential to dry out the code basis, especially when calling the API. Functionality needed in many functions, i.e. API calls and validation, could be separated and stored in a `utils.R` file.

#### Contributing

* A Code of Conduct as a community guideline should be added according to the rOpenSci guidelines. `devtools::use_code_of_conduct()` makes it easy to add the Contributor Covenant template.

#### R CMD check

There was one NOTE when running `R CMD check` locally:

```
* checking dependencies in R code ... NOTE
Namespace in Imports field not imported from: ‘curl’
```

In summary, the package is a very good start, but could be improved by better aligning the code basis and documentation to the rOpenSci guidelines.

Please let me know if something is unclear or doesn't make sense. I am happy to help!
