### 2016-11-02 #78 @perrystephenson
Thanks @njahn82, I appreciate the detailed feedback! I probably won't get a chance to make changes right away, but I'll make sure that all of your feedback is addressed in the next version of the package.
