### 2016-11-03 #78 @kbenoit
Sorry folks, I have been in transit and under the gun for a deadline that passed Oct 31. I will do my best to complete my review by 4 November (Friday).
