### 2016-11-04 #78 @kbenoit
>##### Paper (for packages co-submitting to JOSS)
  >
  >The package contains a `paper.md` with:
  >
  >- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly staing problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

  ---

This is a very specific package aimed at one thing: making it easy for researchers to pull data from the 2014 UK Research Excellence Framework API, for subsequent analysis.  The package wraps API calls detailed on the [REF website](http://impact.ref.ac.uk/CaseStudies/APIhelp.aspx) and returns conveniently classed R objects, for direct analysis in R.

Installed easily, passed full check with `--as-cran` toggled on my system.  (I did not experience the NOTE about `curl` described by the other reviewer.)

The test coverage looks quite impressive (98%), although many or all of these tests could be made much more robust.   For instance, the single test in `test_tag_types.R` is:
```r
test_that("get_tag_types() returns a tibble", {
  # skip_on_cran()
  expect_equal(dim(get_tag_types()),c(13,2))
})
```
which not only does not test whether the return is a tibble, but also only matches the dimensions known in advance.  A more robust test might also compare the values returned to the known tags from the REF website, or that the return type is in fact a tibble class object.

This is more than just being picky, since all of the ability to make sure the package is actually working are done in the `testthat` tests: Every function's **Examples** section is not run.  I realize it's difficult to get API calls to run in tests, but it might be possible to run some of them, but tag them as `\donttest{}` instead of `\dontrun{}`.

#### Adherence to community guidelines

Here I will suggest a few easy-to-make changes:

*  In `DESCRIPTION`, `URL` and `BugReports` are present, but the `Maintainer` field is missing.

*  No `CONTRIBUTING` or `CONDUCT.md` files are present.

#### Adherence to ROpenSci packaging guidelines

*  `NEWS.md`:  This file has no details besides declaring the "initial release", although there are two tagged releases in the GitHub account.  Could link the release announcement in `NEWS.md` to the v1.0 release tag.

*  A `README.Rmd` would better meet the guidelines, including a pre-commit hook to make sure that it gets made into `README.md`.  (See `?devtools::use_readme_rmd`.)

*  This is minor, but the naming could follow the `object_verb` naming guidelines suggested in [`rOpenSci Packaging Guide`](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#funvar), possibly:  `casestudies_get`, `institutions_get`, etc., or even `ref_casestudies_get`, `ref_institutions_get`, ... to make it clear that these are all from this package.  This makes it more like the **stringi** example in the packaging guidelines.  But for such a short package this is really a matter of preference than any sort of requirement.  (But since rOpenSci is designed to keep the R jungle tidy, why not start with even the smallest vines.)

*  To be really picky: According to the guidelines,  `knitr` and `roxygen2` should be added to `Suggests:` in the `DESCRIPTION` file.

This is where the package is weakest, since the documentation is limited to the barest of documentation for the functions.  Even those do not describe much about the context of the options or inputs: for instance `get_case_studies()` did not really explain what the `ID` is or why I might use it instead of a `UKPRN`.  I suggest that this could be vastly expanded in a **Details** section and added as an overview to `refimpact-package.Rd`.  I had to do considerable testing and reading from the UK REF website before I understood what were the inputs to the functions.

Of course a **vignette** is what this package is really crying out for, to show not only how to use it but *why*, including some analysis of the data to show the practical motivation for querying this data through an R package.  This will also get around the problem of not being to run all of the examples live because they involve remote API calls.

#### Motivation

As someone who was actually an individual submitted in this exercise -- although not for a case study -- from one of the listed institutions, and personally involved in managing the staff submitted from my institution, I held the first impression that this would be a really cool package to have, for accessing this data.  However the more I experimented with the package, the more I wondered why the API approach was needed.  The data is static, so the only reasons not to package the data are copyright and size.  Most of the information (except some of the case studies - see http://impact.ref.ac.uk/CaseStudies/Terms.aspx) is governed by a CC license, and so could easily be packaged as data.  The only objection to size applies to the case studies themselves, but again, if the documentation or `README.md` had more on the motivation and/or documentation, I would have a better idea of just how large this is (and whether this size makes it something that is not better simply provided as a flattened large data.frame or "tibble").

The following static tables from the API are CC licensed and could easily be packaged as built-in objects:

*  `institutions`: This table is 155 x 5 data.frame of 20.8k in size
*  `units_of_assessment`:  36 x 3
*  `tag_types`:  13 x 2
*  `values`:  This is much larger but the entire table could be flattened in a way that links to **tag_types**, if we are willing to strongly suspend the principles of relational data normalization (something most users may not know or care about).

This seems to gut the functions from the package, since it leaves only `get_case_studies()`, which might be appropriately handled through an API call.  But here I suggest the package could really enhance value by **adding data-handling functions** that link the static data objects to the structure of what `get_case_studies()` returns, such as ways to flatten the lists that are elements of the return objects from that function.  For instance, the return object from `get_case_studies(ID = c(27,29))` is a 2 x 19 element tibble, but several of those columns (e.g. `Continient`) are variable length lists.  Many users who are not experts in dissecting R objects are going to have trouble with the nesting of lists within data.frames.

In addition, by having the smaller objects as built-in data, the inputs to `get_case_studies()` can be checked for valid values, rather than relying on the API to reject a non-existent ID, for instance.
