### 2016-11-04 #78 @sckott
thanks very much for the review @kbenoit !

@perrystephenson  both reviews in now. let us know if you have any questions. continue discussion here with me and reviewers.
