### 2016-11-06 #78 @perrystephenson
Thanks everyone for the considerable effort you have put in to reviewing my work - I can see why feedback is the first benefit listed in the onboarding README! I've never had such a thorough review of my code before, so I really appreciate all the advice.

I've created a whole bunch of issues in the [refimpact](https://github.com/perrystephenson/refimpact/issues) repository, and I'll start working through them when I get a chance. I'll probably have a few questions (mainly around the benefits of a split data/API package) but I'll wait until I've had a chance to think about the issues more fully before I do.

Thanks again, and I'll let you know how I go!
