### 2017-01-17 #78 @perrystephenson
Hi Scott, sorry I've been a little derailed looking after a new puppy :) I'm planning to get back to this next week, and I'll hopefully have it ready for another review by the end of the month.

Sent from my iPhone

> On 17 Jan 2017, at 7:17 am, Scott Chamberlain <notifications@github.com> wrote:
>
> @perrystephenson let me know if I can help the process move along
>
