### 2017-02-04 perrystephenson/refimpact#20 @perrystephenson
I had a good think about this, and I decided that it was safest to keep the package as a "pure" API wrapper rather than bundling the smaller tables in the package. This is ensure that any changes to the database (as unlikely as that may be) are reflected for users of the package, and also to make the package internally consistent. The case studies tables are many times larger than what would be acceptable in a CRAN package, which means that the API is required for at least that function.

The "flattened tables" idea is a good one, and I will have a look at creating some helper functions to help users navigate the dataset.
