### 2017-02-04 #78 @perrystephenson
I've missed the self-imposed end of January deadline, but I've made fairly significant process working through the issues raised above. You can see how I'm tracking against the issues here:
https://github.com/perrystephenson/refimpact/issues.
