### 2017-03-02 #78 @perrystephenson
I've finished re-writing the package based on your excellent feedback - I don't think there is a single line of code remaining from the original package you reviewed! This has necessarily involved a breaking change to the function syntax (there is now only one user-facing function, named ref_get), so the version number has been incremented to v1.0.0.

I have a couple of notes - excuses if you will!
* There are a few lines without unit test coverage. These lines deal with API failures, which are quite difficult to test. I've tested them locally by removing my network connection, so I am confident that these lines are performing as expected. Test coverage is otherwise more thorough than in the previous version, and I have made use of expect_equal_to_reference() to perform efficient and complete testing of the API methods.
* There is a NOTE in R CMD check about UTF-8 characters in the bundled dataset. I decided that it was better to leave the NOTE there than to replace those UTF-8 characters with ASCII, as they form part of names (such as institutions) in the ref_tags dataset, and changing these characters would overall be detrimental for users. I hope CRAN will agree!

I am happy to go through the changes in more detail, but given that one of the critiques was that the package was hard to use for new users, I think it might be best if I leave you to your own devices with the documentation in the package. Hopefully you will find it far more helpful than before!

Happy to answer any questions, and make any more changes that you suggest. When you're happy with the package, I will tag the v1.0.0 release to align with NEWS.md.

Thanks again for all of your time!
