### 2017-03-02 #78 @masalmon
@perrystephenson Regarding the encoding NOTE, see [Julia Silge's post](http://juliasilge.com/blog/How-I-Stopped/) and her [CRAN comments](https://github.com/juliasilge/janeaustenr/blob/164276c945c0657e776965d7d08034b14a7f53d7/cran-comments.md)  (her package was accepted with the NOTE)
