### 2017-03-02 #78 @kbenoit
I would not worry about the NOTE for the non-ASCII data. **quanteda** has had these since the beginning. Brian Ripley gave the beginnings of a grumble about it, once, but as long as you have
```yaml
Encoding: UTF-8
```
in `DESCRIPTION` then don't worry about it.
