### 2017-03-13 #78 @kbenoit
Sorry it's taken a while to comment on the revision...  I think the update is greatly improved. The vignette is a welcome addition and the tests significantly improved. As I mention above, don't worry about the UTF-8 note.

I still think you could include a lot of the smaller bits of data that are currently accessed through API calls, namely `institutions`, `units_of_assessment`, and `tag_types`. But it works well enough to query these too.

I also still think you could (easily!)  add more description to the `?refimpact` page.

Finally, there are some cool new features in **roxygen2** 6.0 for documentation that you could use, esp for documenting data and packages. Not at all important for the approval process but definitely worth checking out!
