### 2017-03-20 #78 @njahn82
Hi @perrystephenson

Apologies for my late review of your thorough package revision. I am very happy to see that `httr` is now implemented for interacting with the web API and that both vignette and README  guide users through the package.

Here are some some follow-up points after having time to look into your revision:

### Ensure backward compatibility

As introduced, there is now only one function that can be used for calling the API. However, I am afraid that the removed functions are not marked as `.Deprecated`. This is important to ensure backward compatibility with the existing CRAN release. Hadley Wickhams R packages book contains a chapter describing how to deprecate functions:

http://r-pkgs.had.co.nz/release.html

### Function input

I wonder if it would be possible to split up the query parameters, so that I can provide something like

```r
ref_get("SearchCaseStudies", id = "2", tags = "dkkdk")
```

instead of a list object.

I would still not allow numeric input for identifying the studies. In my opinion, it is best practice that IDs should be of class character.

### Simplify nested lists?

I second @kbenoit comment that learners and non-frequent users of R are often unfamiliar with nested lists. The vignette  gives good guidance on how to select list elements. However, it would be great to have at least one parser that additionally returns a subset of data as a tidy `data.frame`. This parser could become a great time saver for users simply wishing to get an overview over the data and to export this selection.

### Vignette and README

Vignette and README contribute greatly to a better understanding of the aim and the usage of this package. Especially, the benefits and limits of the API implementation are very well described. One API problem, no call for retrieving all datasets at once, is very well solved by showing how to extract the entire dataset. There is some potential to shorten the example by using `lapply` instead of a for-loop:

``` r
# load dev version
devtools::install_github("perrystephenson/refimpact")
#> Skipping install of 'refimpact' from a github remote, the SHA1 (a1ccaafb) has not changed since last install.
#>   Use `force = TRUE` to force installation
# get IDs
uoa_table <- refimpact::ref_get("ListUnitsOfAssessment")
# apply 'ref_get' over all IDs
tt <- lapply(uoa_table$ID, function(x) refimpact::ref_get("SearchCaseStudies",
  query = list(UoA = x)))
```

Please provide at least one usage example in the README.

Finally, I wonder if you want to include your vignette as part of the README. If so, simply add

<code>
</code>

to your `README.Rmd`.
