### 2017-03-26 #78 @perrystephenson
Thanks everyone,

I'll try and respond to these in order.

@kbenoit

I've made comments about the bundling of the smaller tables [here](https://github.com/perrystephenson/refimpact/issues/20#issuecomment-277427290). If the case studies were small enough to bundle with the package then I would have just gone with a 100% data package, but given that users must use the API to make use of the package, there isn't any harm in getting the users to pull down the small tables. The possible benefit is that if there are new records added to the database, then users will get these records.

I have raised an issue for the help ?refimpact documentation and will sort it out soon.

And I'll take a look at Roxygen2 6.0!

@njahn82

I thought about backwards compatibility, and my thinking was along the lines of:

1. It has barely any downloads (672 from the RStudio CRAN mirror as of today), and a Google search for library(refimpact) didn't return any meaningful hits. I doubt anyone is going to notice the lack of backward compatibility.
2. It has no reverse dependencies on CRAN
3. The old function names were not compliant with rOpenSci guidelines
4. I incremented the major version number, which indicates breaking changes

I'm happy to take guidance on this if you think backwards compatibility is worth maintaining!

I can definitely split up the list - I left it as a list partly due to convenience (httr takes a list) and partly due to having seen that pattern in many other packages (and in fact in lots of base R functions). Is there a reason you think it's better as individual arguments?

Regarding the simplification of nested lists into useful data structures, I had a bit of a look at what this would take, and there is a fair bit of work in it. I've left the issue open and tagged it as something to look at for v1.1, unless you think it's important enough that it needs to be part of v1.0. I suspect it's going to take a while to think through all of the use cases though.

I'll have another look at the vignette and README based on your comments - I might use purrr to shorten the code example above, as I don't think that the apply functions are particularly good at conveying meaning to users looking at vignettes, and purrr tends to be a little easier to read.
