### 2017-03-27 #78 @njahn82
Thanks for your reply, @perrystephenson

I am feeling unsure about what policies and good practices concerning deprecated or defunct functions are in place at ropensci, perhaps @sckott can help? Maybe there are users of your package who would appreciate it to find guidance about changes.

Regarding the `ref_get()` input, I found it a good practice among web api packages to call functions with simple atomic vectors because they make it easier to grasp what input is required. In the case of `ref_get()` I had to check with the API documentation to get an idea about what search parameters exist and what they mean. So I want to suggest to at least expand the `ref_get()` documentation with examples that cover the five existing query parameters.

It is great that you are planning to provide some parsers, and if you would like to add them in another release I am ok with it.

Looking forward to find purrr examples in the vignette!
