### 2017-03-30 #78 @sckott
In terms of deprecating and defunct, since it's already on CRAN, i advise to make a function deprecated in the next pkg version, then version after that make the function defunct - So that users have one version with warning about function going away, then next version it's gone - And do maintain man files for deprecated and defunct functions
