### 2017-06-29 #78 @perrystephenson
Hi @sckott and sorry for the delay. Following from the discussion above:

* I've restored all of the functions from v0.1.0 and added a deprecation message to those functions (and made an appropriate note in the NEWS.md file)
* Made a minor addition to ?refimpact (pointing the reader to the vignette - no point in duplicating the vignette in ?refimpact)
* Added lots of detail and examples to ?ref_get

I've decided to leave the "get all tags" example in the vignette as a loop. I think that loops are likely to be more readily understood by users who are new to programming (and new to R), which makes it a bit more friendly than using an apply function or purrr.

These changes have all been pushed to GitHub, so feel free to take a look.

Thanks again for the feedback!
