### 2017-06-30 #78 @sckott
thanks @perrystephenson - vacation for a week now, will have a look when I get back, sorry for the delay!
