### 2017-07-11 #78 @sckott
@perrystephenson looking good, checked it over, and nearly all is good. Just one thing

here https://github.com/perrystephenson/refimpact/blob/master/R/ref_get.R#L120 you are parsing content and then parsing JSON, whereas the HTTP status code check happens after that - why not do the status code check directly after https://github.com/perrystephenson/refimpact/blob/master/R/ref_get.R#L109
