### 2017-07-12 #78 @perrystephenson
Hi @sckott excellent question. The reason for this ordering is that the API gives useful error messages for some error conditions, and those error messages are worth passing on to the user. So following the API call I firstly check whether the response is a JSON (and throw an error if not), then I parse the JSON (which I now know I can do safely) to get either the content or the error message, then if there is an error I can add the parsed error message to the error string which gets presented to the user.

The alternative is to make the error messages less useful, which I think would be an inferior outcome.
