### 2017-07-14 #78 @sckott
True, you do check for JSON first, then parse after that is confirmed, so seems okay.  - Def. agree to give useful error messages whenever possible
