### 2017-07-15 #78 @perrystephenson
Thanks @sckott!

I've done all of those tasks now - is there anything I need to do before submitting this new version to CRAN? Do I need to add any rOpenSci references to DESCRIPTION?

Once again, thanks to everyone for the incredibly detailed feedback - I have learned a lot through the process! I really appreciate the time and effort you've committed to this process.
