### 2017-07-15 #78 @sckott
>  is there anything I need to do before submitting this new version to CRAN? Do I need to add any rOpenSci references to DESCRIPTION?

Just make sure to update the URLs in your DESCRIPTION file , right now they have the old github username

thanks for your submission!
